search: extract bloom_walk, callee_query, scope from relational queries
Pull three helpers out of `callers.rs` / `callees.rs`:

- `bloom_walk::read_with_bloom_check` — shared read + size-gate + bloom
  prefilter, called by `find_callers_batch` and the import-resolution
  loop in `resolve_callees` (was duplicated inline in both).
- `callee_query` — per-language callee query strings + compiled-Query
  cache, no longer reached across modules via
  `super::callees::callee_query_str` from `callers.rs`.
- `scope` — `EnclosingScope` and `enclosing_definition_at`, whose only
  consumer is `search/mod.rs`'s formatter; previously housed in
  `callers.rs` by accident of geometry.

Net +38 prod / +61 test LOC: the bloom_walk helper's fixed overhead
exceeds the dedup savings, and the move targets carry their tests.
The win is cohesion, not size — `callers.rs` 902→536, `callees.rs`
690→526; cross-module `pub(super)`/`pub(crate)` reaches are gone;
each new module has one concern. 315/315 tests green.

(cherry picked from commit e3c490766ecb4dea35f3bff011e78dcafb3f3959)
